requirements.txt
# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# Database
# =============================================================================
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# =============================================================================
# Redis & Background Tasks
# =============================================================================
redis>=5.0.0
taskiq>=0.11.0
taskiq-redis>=1.0.0

# =============================================================================
# HTTP Client
# =============================================================================
httpx>=0.26.0

# =============================================================================
# Security
# =============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# =============================================================================
# Logging & Observability
# =============================================================================
structlog>=24.1.0

# =============================================================================
# Configuration
# =============================================================================
pyyaml>=6.0.1
python-dotenv>=1.0.0

# =============================================================================
# CLI
# =============================================================================
click>=8.1.0
rich>=13.7.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0

# =============================================================================
# Code Quality
# =============================================================================
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0
